+++ bin/cli/src/validate/mod.rs
@@ -23,8 +23,8 @@ use kailua_build::KAILUA_FPVM_ID;
 use kailua_common::config::config_hash;
 use kailua_common::journal::ProofJournal;
 use kailua_common::precondition::PreconditionValidationData;
-use kailua_host::channel::AsyncChannel;
 use kailua_prover::backends::boundless::BoundlessArgs;
+use kailua_prover::channel::AsyncChannel;
 use kailua_prover::proof::{proof_file_name, read_proof_file};
 use kailua_sync::args::parse_address;
 use kailua_sync::args::SyncArgs;
