+++ crates/prover/Cargo.toml
@@ -12,6 +12,7 @@ categories.workspace = true
 
 [dependencies]
 anyhow.workspace = true
+async-channel.workspace = true
 async-trait.workspace = true
 bincode.workspace = true
 bytemuck.workspace = true
@@ -20,6 +21,8 @@ clap.workspace = true
 hex.workspace = true
 human_bytes.workspace = true
 rkyv.workspace = true
+serde_json.workspace = true
+tempfile.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
@@ -34,6 +37,9 @@ kailua-sync.workspace = true
 
 kona-cli.workspace = true
 kona-derive.workspace = true
+kona-genesis.workspace = true
+kona-host.workspace = true
+kona-registry.workspace = true
 kona-preimage.workspace = true
 kona-proof.workspace = true
 kona-protocol.workspace = true
