+++ crates/prover/src/backends/boundless.rs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{ProvingArgs, ProvingError};
+use crate::args::ProvingArgs;
+use crate::ProvingError;
 use alloy::network::Ethereum;
 use alloy::transports::http::reqwest::Url;
 use alloy_primitives::utils::parse_ether;
