+++ crates/prover/src/client/mod.rs
@@ -15,6 +15,7 @@
 use kona_preimage::{HintWriter, OracleReader};
 use kona_std_fpvm::{FileChannel, FileDescriptor};
 
+pub mod native;
 pub mod proving;
 pub mod witgen;
