+++ crates/prover/src/config.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::args::KailuaHostArgs;
+use crate::args::ProveArgs;
 use kailua_sync::provider::optimism::fetch_rollup_config;
 use kona_genesis::RollupConfig;
 use kona_registry::Registry;
@@ -20,18 +20,18 @@ use tempfile::TempDir;
 use tokio::fs;
 use tracing::{debug, info};
 
-pub async fn generate_rollup_config(
-    cfg: &mut KailuaHostArgs,
+pub async fn generate_rollup_config_file(
+    args: &mut ProveArgs,
     tmp_dir: &TempDir,
 ) -> anyhow::Result<RollupConfig> {
     // generate a RollupConfig for the target network
-    Ok(match cfg.kona.read_rollup_config().ok() {
+    Ok(match args.kona.read_rollup_config().ok() {
         Some(rollup_config) => rollup_config,
         None => {
             let registry = Registry::from_chain_list();
             let tmp_cfg_file = tmp_dir.path().join("rollup-config.json");
-            if let Some(rollup_config) = cfg.kona.l2_chain_id.and_then(|chain_id| {
-                if cfg.bypass_chain_registry {
+            if let Some(rollup_config) = args.kona.l2_chain_id.and_then(|chain_id| {
+                if args.bypass_chain_registry {
                     None
                 } else {
                     registry.rollup_configs.get(&chain_id)
@@ -46,8 +46,8 @@ pub async fn generate_rollup_config(
             } else {
                 info!("Fetching rollup config from nodes.");
                 fetch_rollup_config(
-                    cfg.op_node_address.as_ref().unwrap().as_str(),
-                    cfg.kona
+                    args.op_node_address.as_ref().unwrap().as_str(),
+                    args.kona
                         .l2_node_address
                         .clone()
                         .expect("Missing l2-node-address")
@@ -56,9 +56,9 @@ pub async fn generate_rollup_config(
                 )
                 .await?;
             }
-            cfg.kona.rollup_config_path = Some(tmp_cfg_file);
-            debug!("{:?}", cfg.kona.rollup_config_path);
-            cfg.kona.read_rollup_config()?
+            args.kona.rollup_config_path = Some(tmp_cfg_file);
+            debug!("{:?}", args.kona.rollup_config_path);
+            args.kona.read_rollup_config()?
         }
     })
 }
